evaluate equity onClick #9
