+++ src/App.css
@@ -34,3 +34,13 @@ body {
 .margin-left {
     margin-left : 30px;
 }
+
+#eval-btn {
+    background-color: #6b94d6;
+    border-color: #6b94d6;
+}
+
+#clear-btn {
+    background-color: #de6360;
+    border-color: #de6360;
+}
