+++ src/equitycalculator/components/buttons/clear.jsx
@@ -7,6 +7,7 @@ export const ClearButton = (props) => {
 
     return (
         <button
+            id="clear-btn"
             className="btn btn-danger margin-top margin-left"
             onClick={props.onClick}>
             Clear
