+++ src/equitycalculator/components/buttons/evaluate.jsx
@@ -1,6 +1,9 @@
 import * as React from 'react'
 
-export const EvaluateButton = () =>
-    <button className="btn btn-success margin-top">
+export const EvaluateButton = (props) =>
+    <button
+        className="btn btn-success margin-top"
+        id="eval-btn"
+        onClick={() => props.saveEquity(props.slots, props.savedEquities)}>
         Evaluate
     </button>
